# openAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Freya Salon Backend API",
    "description": "\n    ## Freya Beauty Salon Management System API\n    \n    Bu API Freya go'zallik saloni boshqaruv tizimi uchun yaratilgan.\n    \n    ### Asosiy funksiyalar:\n    - 🔐 **Autentifikatsiya va avtorizatsiya** (JWT token)\n    - 👥 **Foydalanuvchilar boshqaruvi** (Admin, User, Employee)\n    - 🏢 **Salon boshqaruvi** (Salon ma'lumotlari, xizmatlar)\n    - 📅 **Vaqt jadvali boshqaruvi** (Booking, Schedule)\n    - 💳 **To'lov tizimi** (Click.uz integratsiyasi)\n    - 📱 **SMS xabarnomalar** (Eskiz.uz integratsiyasi)\n    - 🌐 **Tarjima xizmati** (DeepL API)\n    \n    ### Texnologiyalar:\n    - **FastAPI** - Web framework\n    - **SQLAlchemy** - ORM\n    - **PostgreSQL/SQLite** - Database\n    - **JWT** - Authentication\n    - **Pydantic** - Data validation\n    \n    ### Versiya: 2.0.0\n    ",
    "contact": { "name": "Freya Development Team", "email": "dev@freya.uz" },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.0.0"
  },
  "servers": [
    { "url": "http://localhost:8000", "description": "Development server" },
    {
      "url": "https://freya-salon-backend.herokuapp.com",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/api/sms/send-verification-code": {
      "post": {
        "tags": ["SMS"],
        "summary": "Tasdiqlash kodini yuborish",
        "description": "Telefon raqamini tasdiqlash uchun kod yuborish\n\n- **phone**: Telefon raqami (+998901234567 formatida)",
        "operationId": "send_verification_code_api_sms_send_verification_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendVerificationCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Send Verification Code Api Sms Send Verification Code Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/sms/send-reset-password-code": {
      "post": {
        "tags": ["SMS"],
        "summary": "Parolni tiklash kodini yuborish",
        "description": "Parolni tiklash uchun kod yuborish\n\n- **phone**: Telefon raqami (+998901234567 formatida)",
        "operationId": "send_reset_password_code_api_sms_send_reset_password_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendResetPasswordCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Send Reset Password Code Api Sms Send Reset Password Code Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/sms/send-change-phone-code": {
      "post": {
        "tags": ["SMS"],
        "summary": "Telefon raqamini o'zgartirish kodini yuborish",
        "description": "Telefon raqamini o'zgartirish uchun kod yuborish\n\n- **old_phone**: Eski telefon raqami\n- **new_phone**: Yangi telefon raqami",
        "operationId": "send_change_phone_code_api_sms_send_change_phone_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendChangePhoneCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Send Change Phone Code Api Sms Send Change Phone Code Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/sms/send-registration-code": {
      "post": {
        "tags": ["SMS"],
        "summary": "Ro'yxatdan o'tish kodini yuborish",
        "description": "Ro'yxatdan o'tish uchun kod yuborish\n\n- **phone**: Telefon raqami (+998901234567 formatida)",
        "operationId": "send_registration_code_api_sms_send_registration_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendRegistrationCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Send Registration Code Api Sms Send Registration Code Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/sms/send-payment-card-code": {
      "post": {
        "tags": ["SMS"],
        "summary": "To'lov kartasi tasdiqlash kodini yuborish",
        "description": "To'lov kartasini tasdiqlash uchun kod yuborish\n\n- **phone**: Telefon raqami\n- **card_number**: Karta raqami",
        "operationId": "send_payment_card_code_api_sms_send_payment_card_code_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendPaymentCardCodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Send Payment Card Code Api Sms Send Payment Card Code Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/sms/balance": {
      "get": {
        "tags": ["SMS"],
        "summary": "SMS balansini tekshirish",
        "description": "Eskiz.uz hisobidagi SMS balansini tekshirish",
        "operationId": "check_sms_balance_api_sms_balance_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Check Sms Balance Api Sms Balance Get"
                }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/sms/status/{message_id}": {
      "get": {
        "tags": ["SMS"],
        "summary": "SMS holatini tekshirish",
        "description": "Yuborilgan SMS holatini tekshirish\n\n- **message_id**: SMS ID",
        "operationId": "check_sms_status_api_sms_status__message_id__get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "message_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Message Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Check Sms Status Api Sms Status  Message Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/sms/refresh-token": {
      "post": {
        "tags": ["SMS"],
        "summary": "Eskiz.uz tokenini yangilash",
        "description": "Eskiz.uz access tokenini yangilash",
        "operationId": "refresh_eskiz_token_api_sms_refresh_token_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Refresh Eskiz Token Api Sms Refresh Token Post"
                }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SendChangePhoneCodeRequest": {
        "properties": {
          "old_phone": { "type": "string", "title": "Old Phone" },
          "new_phone": { "type": "string", "title": "New Phone" }
        },
        "type": "object",
        "required": ["old_phone", "new_phone"],
        "title": "SendChangePhoneCodeRequest"
      },
      "SendPaymentCardCodeRequest": {
        "properties": {
          "phone": { "type": "string", "title": "Phone" },
          "card_number": { "type": "string", "title": "Card Number" }
        },
        "type": "object",
        "required": ["phone", "card_number"],
        "title": "SendPaymentCardCodeRequest"
      },
      "SendRegistrationCodeRequest": {
        "properties": { "phone": { "type": "string", "title": "Phone" } },
        "type": "object",
        "required": ["phone"],
        "title": "SendRegistrationCodeRequest"
      },
      "SendResetPasswordCodeRequest": {
        "properties": { "phone": { "type": "string", "title": "Phone" } },
        "type": "object",
        "required": ["phone"],
        "title": "SendResetPasswordCodeRequest"
      },
...
      "SendVerificationCodeRequest": {
        "properties": { "phone": { "type": "string", "title": "Phone" } },
        "type": "object",
        "required": ["phone"],
        "title": "SendVerificationCodeRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": { "HTTPBearer": { "type": "http", "scheme": "bearer" } }
  }
}